Move stats to junit project
