+++ docker-compose-rule-junit4/src/main/java/com/palantir/docker/compose/stats/ContainerStats.java
@@ -1,0 +1,38 @@
+/*
+ * (c) Copyright 2019 Palantir Technologies Inc. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.palantir.docker.compose.stats;
+
+import com.palantir.docker.compose.stats.ImmutableContainerStats;
+import java.time.Duration;
+import java.util.Optional;
+import org.immutables.value.Value;
+
+@Value.Immutable
+public interface ContainerStats {
+    String containerName();
+    Optional<Duration> timeTakenToBecomeHealthy();
+
+    default boolean startedSuccessfully() {
+        return timeTakenToBecomeHealthy().isPresent();
+    }
+
+    class Builder extends ImmutableContainerStats.Builder { }
+
+    static Builder builder() {
+        return new Builder();
+    }
+}
