+++ docker-compose-rule-junit4/src/main/java/com/palantir/docker/compose/stats/Stats.java
